don't overload public body categories when testing, in order to break tests

+++ lib/alavetelitheme.rb
@@ -8,9 +8,11 @@ end
 
 # In order to have the theme lib/ folder ahead of the main app one,
 # inspired in Ruby Guides explanation: http://guides.rubyonrails.org/plugins.html
-%w{ lib }.each do |dir|
-  path = File.join(File.dirname(__FILE__), dir)
-  $LOAD_PATH.insert(0, path)
-  ActiveSupport::Dependencies.autoload_paths << path
-  ActiveSupport::Dependencies.autoload_once_paths.delete(path)
+if ENV["RAILS_ENV"] != "test"
+    %w{ lib }.each do |dir|
+      path = File.join(File.dirname(__FILE__), dir)
+      $LOAD_PATH.insert(0, path)
+      ActiveSupport::Dependencies.autoload_paths << path
+      ActiveSupport::Dependencies.autoload_once_paths.delete(path)
+    end
 end
